Refactors theme and baseController into new base module
