--- app/src/main/java/io/github/koss/mammut/feature/instance/subfeature/navigation/ModuleOwner.kt
@@ -1,8 +1,0 @@
-package io.github.koss.mammut.feature.instance.subfeature.navigation
-
-interface ModuleOwner {
-
-    fun stashModule(key: String, module: Any)
-
-    fun retrieveModule(key: String): Any
-}
