+++ app/src/main/java/io/github/koss/mammut/feature/joininstance/JoinInstanceActivity.kt
@@ -17,7 +17,6 @@ import io.github.koss.mammut.extension.applicationComponent
 import io.github.koss.mammut.extension.observe
 import io.github.koss.mammut.extension.provideViewModel
 import io.github.koss.mammut.extension.snackbar
-import io.github.koss.mammut.feature.base.BaseActivity
 import io.github.koss.mammut.feature.base.InputError
 import io.github.koss.mammut.feature.instancebrowser.InstanceBrowserActivity
 import io.github.koss.mammut.feature.joininstance.dagger.JoinInstanceModule
@@ -29,6 +28,7 @@ import org.jetbrains.anko.startActivity
 import javax.inject.Inject
 import androidx.browser.customtabs.CustomTabsIntent
 import io.github.koss.mammut.R
+import io.github.koss.mammut.base.BaseActivity
 import org.jetbrains.anko.colorAttr
 import saschpe.android.customtabs.CustomTabsHelper
 import saschpe.android.customtabs.WebViewFallback
