+++ app/src/main/java/io/github/koss/mammut/feature/settings/SettingsController.kt
@@ -15,18 +15,18 @@ import com.github.ajalt.flexadapter.FlexAdapter
 import com.github.ajalt.flexadapter.register
 import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
 import io.github.koss.mammut.R
+import io.github.koss.mammut.base.BaseController
+import io.github.koss.mammut.base.themes.ThemeEngine
 import io.github.koss.mammut.component.retention.retained
 import io.github.koss.mammut.dagger.MammutViewModelFactory
 import io.github.koss.mammut.dagger.application.ApplicationScope
 import io.github.koss.mammut.extension.observe
 import io.github.koss.mammut.extension.startActivity
 import io.github.koss.mammut.feature.instance.InstanceActivity
-import io.github.koss.mammut.feature.instance.subfeature.navigation.BaseController
 import io.github.koss.mammut.feature.instancebrowser.InstanceBrowserActivity
 import io.github.koss.mammut.feature.settings.dagger.SettingsModule
 import io.github.koss.mammut.feature.settings.dagger.SettingsScope
 import io.github.koss.mammut.feature.settings.model.*
-import io.github.koss.mammut.feature.themes.ThemeEngine
 import kotlinx.android.extensions.CacheImplementation
 import kotlinx.android.extensions.ContainerOptions
 import kotlinx.android.synthetic.main.controller_settings.*
