+++ base/src/main/java/io/github/koss/mammut/base/BaseActivity.kt
@@ -1,18 +1,14 @@
-package io.github.koss.mammut.feature.base
+package io.github.koss.mammut.base
 
 import android.content.Context
 import android.os.Build
 import android.os.Bundle
 import android.view.View
-import android.view.View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR
 import androidx.annotation.CallSuper
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
-import io.github.koss.mammut.R
-import io.github.koss.mammut.feature.themes.ThemeEngine
+import io.github.koss.mammut.base.themes.ThemeEngine
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.GlobalScope
-import org.jetbrains.anko.attr
 import org.jetbrains.anko.colorAttr
 import org.jetbrains.anko.contentView
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper
@@ -41,14 +37,14 @@ abstract class BaseActivity: AppCompatActivity(), CoroutineScope by GlobalScope 
                 systemUiVisibility = when {
                     themeEngine.isLightTheme -> View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR.run {
                         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                            this or SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR
+                            this or View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR
                         } else {
                             this
                         }
                     }
                     else -> systemUiVisibility and (View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR.run {
                         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                            this or SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR
+                            this or View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR
                         } else {
                             this
                         }
