+++ base/src/main/java/io/github/koss/mammut/base/themes/ThemeConfig.kt
@@ -1,0 +1,6 @@
+package io.github.koss.mammut.base.themes
+
+interface ThemeConfig {
+
+    var currentThemeId: String?
+}
