+++ base/src/main/res/values/theme_standard.xml
