+++ base/src/main/res/values/theme_standard_light.xml
