+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':instances', ':toot', ':data'
+include ':app', ':instances', ':toot', ':data', ':base'
